ZXP-159 Move NotificationService in services folder
