+++ front-angular/src/app/services/index.ts
@@ -3,4 +3,4 @@ export * from './company.service';
 export * from './user.service';
 export * from './avis.service';
 export * from './offer-view.service';
-
+export * from './notifications.service';
